api/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pydantic
pandas
xgboost
geopy
transformers
torch
huggingface_hub
python-dotenv
google-generativeai